+++ vlaanderen/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Removed option to create Templates - [pr](https://agiv.visualstudio.com/Metadata/_git/MetadataGeonetwork/pullrequest/49947?_a=files)
 - Prevent workflow activation for harvested records - [pr](https://agiv.visualstudio.com/Metadata/_git/MetadataGeonetwork/pullrequest/50341)
 - Fixed spurious temporal history entries for usergroups on login - [pr](https://agiv.visualstudio.com/Metadata/_git/MetadataGeonetwork/pullrequest/50479)
+- Display harvested source information and link for harvested records in the record view - [pr](https://agiv.visualstudio.com/Metadata/_git/MetadataGeonetwork/pullrequest/50868)
 
 ### core
 - Added OSLO Datastandaarden thesaurus for use in ISO - [pr](https://agiv.visualstudio.com/Metadata/_git/MetadataGeonetwork/pullrequest/48724) / [pr-dcat](https://github.com/metadata101/dcat-ap/pull/127)